[module3-task1] Fix error on click on filter and separete 'getPoint' and 'getPoints'
